TKW-76 - Refactor imports and fix navigation typo in SettingsScreen
