+++ .rnstorybook/stories/Text.stories.tsx
@@ -1,7 +1,7 @@
 import { View } from "react-native";
 import type { Meta, StoryObj } from "@storybook/react";
 import { action } from "storybook/actions";
-import Text from "../../components/Text";
+import Text from "../../components/UI/Text";
 
 const meta = {
   title: "Components/Text",
