+++ .rnstorybook/storybook.requires.ts
@@ -1,3 +1,4 @@
+/* eslint-disable */
 /* do not change this file, it is auto generated by storybook. */
 import { start, updateView, View } from '@storybook/react-native';
